+++ src/routing/http_wrapper.h
@@ -9,7 +9,7 @@ Copyright (c) 2015-2022, Julien Coupey.
 All rights reserved (see LICENSE).
 
 */
-#include "../include/rapidjson/document.h"
+#include "../include/rapidjson/include/rapidjson/document.h"
 
 #include "routing/wrapper.h"
 #include "structures/typedefs.h"